made the search and text selection pages separate
