+++ src/main/java/wikispeak/CreateAudioScreenController.java
@@ -23,7 +23,6 @@ import java.util.regex.Pattern;
 public class CreateAudioScreenController extends Controller{
 
     @FXML private BorderPane rootBorderPane;
-    @FXML private TextField searchField;
     @FXML private TextArea textOutput;
     @FXML private ComboBox voiceSelection;
     @FXML private Button createAudioButton;
@@ -39,63 +38,27 @@ public class CreateAudioScreenController extends Controller{
         });
         //listens to changes in the selected text
         textOutput.selectedTextProperty().addListener(((observable, oldValue, newValue) -> {
+            //createAudioButton is disabled if: file name field is empty, or no voice is selected, or no text is selected in the text area
             createAudioButton.setDisable(newValue.isEmpty() || voiceSelection.getValue() == null || audioFileNameField.getText().isEmpty());
         }));
         populateVoiceSelectionBox();
+        populateTextArea();
     }
 
     @FXML
-    private void handleMainMenu() throws IOException {
-        switchScenes(rootBorderPane, "MainMenu.fxml");
+    private void handleBackToSearch() throws IOException {
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION, "Any audio files you have created will be deleted.\nAre you sure you want to go back?");
+        alert.getDialogPane().setMinHeight(Region.USE_PREF_SIZE);
+        Optional<ButtonType> result = alert.showAndWait();
+        if (result.isPresent() && result.get() == ButtonType.OK) {
+            //only switch scene after confirmation
+            switchScenes(rootBorderPane, "SearchScreen.fxml");
+        }
     }
 
     @FXML
-    private void handleSearch(){
-
-        String currentSearch = searchField.getText();
-
-        if(!currentSearch.isEmpty()) {
-            infoText.setText("Searching...");
-            //while searching, WHAT SHOULD BE DISABLED?
-            createAudioButton.setDisable(true);
-
-            Thread searchThread = new Thread(new Task<Void>(){
-                @Override
-                protected Void call() throws Exception {
-                    Command wikitCommand = new Command("wikit " + currentSearch + " | sed 's/\\([.!?]\\) \\([[:upper:]]\\)/\\1\\n\\2/g' > .temp_text.txt");
-                    wikitCommand.execute();
-                    return null;
-                }
-                @Override
-                protected void done(){
-                    Platform.runLater(() -> postSearchUpdateGUI());
-                }
-            });
-
-            searchThread.start();
-        }
-    }
-
-    /**
-     * Updates the GUI appropriately based on the outcome of a Wikit Search
-     */
-    private void postSearchUpdateGUI() {
-        Command command = new Command("cat .temp_text.txt | grep -Fwq \":^(\"");
-        if(command.execute() == 0){
-            //nothing found on Wikipedia, update GUI to inform user
-            command = new Command("cat .temp_text.txt");
-            command.execute();
-            textOutput.setText(command.getStream());
-        }
-        else{
-            //term found on wikipedia, show text in textArea
-            command = new Command("cat .temp_text.txt");
-            command.execute();
-            textOutput.setText(command.getStream());
-            //textOutput becomes enabled because search was successful,
-            textOutput.setDisable(false);
-        }
-        infoText.setText("Edit and highlight the text below. Enter a name and voice to create an audio file.");
+    private void handleNext(){
+        System.out.println("Next Pressed");
     }
 
     @FXML
@@ -109,13 +72,12 @@ public class CreateAudioScreenController extends Controller{
 
         String selectedVoice = (String)voiceSelection.getValue();
         String textSelection = textOutput.getSelectedText();
-
         if(!correctTextSelection(textSelection)){
             textSelection = "This is my voice";
         }
-        //read out the selected text in the selected voice
         String finalTextSelection = textSelection;
 
+        //read out the selected text in the selected voice
         Thread previewVoiceThread = new Thread(new Task<Void>(){
             @Override
             protected Void call() throws Exception {
@@ -160,6 +122,27 @@ public class CreateAudioScreenController extends Controller{
         }
     }
 
+
+
+    /**
+     * Adds Festival voices to the VoiceSelectionComboBox
+     */
+    private void populateVoiceSelectionBox(){
+        String[] voiceNameArray = {"kal_diphone", "akl_nz_jdt_diphone", "akl_nz_cw_cg_cg"};
+        for(int i = 0; i < voiceNameArray.length; i++){
+            voiceSelection.getItems().add(voiceNameArray[i]);
+        }
+    }
+
+    /**
+     * Adds the search result text to the text area
+     */
+    private void populateTextArea(){
+        Command command = new Command("cat .temp_text.txt");
+        command.execute();
+        textOutput.setText(command.getStream());
+    }
+
     /**
      * Checks the user has selected the right amount of text to synthesise
      * @return false if no words are selected, or if the selection is larger than 40 words (too many to synthesise)
@@ -182,16 +165,6 @@ public class CreateAudioScreenController extends Controller{
     }
 
     /**
-     * Adds Festival voices to the VoiceSelectionComboBox
-     */
-    private void populateVoiceSelectionBox(){
-        String[] voiceNameArray = {"kal_diphone", "akl_nz_jdt_diphone", "akl_nz_cw_cg_cg"};
-        for(int i = 0; i < voiceNameArray.length; i++){
-            voiceSelection.getItems().add(voiceNameArray[i]);
-        }
-    }
-
-    /**
      * Checks if the given filename doesn't contain forbidden characters
      * @param fileName
      * @return true or false
@@ -208,15 +181,15 @@ public class CreateAudioScreenController extends Controller{
     }
 
     /**
-     * Checks if specified file already exists and, if so, prompts for overwrite
+     * Checks if specified audio file already exists and, if so, prompts for overwrite
      * @param fileName the name of the creation trying to be made
      */
     private boolean canOverwrite(String fileName){
-        File file = new File("audio" + System.getProperty("file.separator") + fileName + ".mp4");
+        File file = new File("audio" + System.getProperty("file.separator") + fileName + ".wav");
         if(file.exists()){
 
-            //create a show an alert
-            Alert alert = new Alert(Alert.AlertType.CONFIRMATION, "Audiofile with name \"" + fileName + "\" already exists. Overwrite?");
+            //show an alert
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION, "Audio file with name \"" + fileName + "\" already exists. Overwrite?");
             alert.getDialogPane().setMinHeight(Region.USE_PREF_SIZE);
             Optional<ButtonType> result = alert.showAndWait();
 
@@ -269,6 +242,23 @@ public class CreateAudioScreenController extends Controller{
             creationThread.start();
         }
     }
+    /**
+     * Deletes specified creation after asking for confirmation
+     * @param creation
+     * @throws IOException
+    private void deleteCreation(String creation) throws IOException {
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION, "Are you sure you want to delete \"" + creation + "\"?");
+        alert.getDialogPane().setMinHeight(Region.USE_PREF_SIZE);
+        Optional<ButtonType> result = alert.showAndWait();
+        if (result.isPresent() && result.get() == ButtonType.OK) {
+            File file = new File("creations" + System.getProperty("file.separator") + creation + ".mp4");
+            file.delete();
+
+            //refresh the scene
+            switchScenes(rootBorderPane, "DeleteScreen.fxml");
+        }
+    }
+
     */
 
 }
