--- src/main/resources/wikispeak/DeleteScreen.fxml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ScrollPane?>
-<?import javafx.scene.layout.BorderPane?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-<BorderPane fx:id="rootBorderPane" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/" xmlns:fx="http://javafx.com/fxml/" fx:controller="wikispeak.DeleteScreenController">
-   <center>
-      <ScrollPane prefHeight="336.0" prefWidth="600.0" BorderPane.alignment="CENTER">
-         <content>
-            <VBox fx:id="creationVBox" prefHeight="261.0" prefWidth="562.0" spacing="5.0">
-               <opaqueInsets>
-                  <Insets />
-               </opaqueInsets></VBox>
-         </content>
-         <BorderPane.margin>
-            <Insets bottom="10.0" left="10.0" right="10.0" />
-         </BorderPane.margin>
-      </ScrollPane>
-   </center>
-   <bottom>
-      <Button minHeight="30.0" minWidth="150.0" mnemonicParsing="false" onAction="#handleMainMenu" text="Back to Main Menu" BorderPane.alignment="CENTER_LEFT" />
-   </bottom>
-   <top>
-      <VBox prefHeight="95.0" prefWidth="600.0" BorderPane.alignment="CENTER">
-         <children>
-            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Delete Existing Creation" textAlignment="CENTER" wrappingWidth="600.0">
-               <font>
-                  <Font size="40.0" />
-               </font>
-               <VBox.margin>
-                  <Insets top="10.0" />
-               </VBox.margin>
-            </Text>
-            <Text fx:id="infoText" strokeType="OUTSIDE" strokeWidth="0.0" text="Select a Creation to Delete...">
-               <VBox.margin>
-                  <Insets left="10.0" top="10.0" />
-               </VBox.margin>
-            </Text>
-         </children>
-      </VBox>
-   </top>
-</BorderPane>
