+++ src/main/resources/wikispeak/MainMenu.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<BorderPane fx:id="rootBorderPane" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/" xmlns:fx="http://javafx.com/fxml/" fx:controller="wikispeak.MainMenuController">
+<BorderPane fx:id="rootBorderPane" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="wikispeak.MainMenuController">
    <center>
       <VBox alignment="CENTER" BorderPane.alignment="CENTER">
          <children>
@@ -18,14 +18,7 @@
                <font>
                   <Font size="18.0" />
                </font></Button>
-            <Button mnemonicParsing="false" onAction="#handlePlay" prefHeight="50.0" prefWidth="400.0" text="Play/View Existing Creations">
-               <VBox.margin>
-                  <Insets bottom="20.0" />
-               </VBox.margin>
-               <font>
-                  <Font size="18.0" />
-               </font></Button>
-            <Button mnemonicParsing="false" onAction="#handleDelete" prefHeight="50.0" prefWidth="400.0" text="Delete Existing Creations">
+            <Button mnemonicParsing="false" onAction="#handlePlay" prefHeight="50.0" prefWidth="400.0" text="Play/Delete Existing Creations">
                <VBox.margin>
                   <Insets bottom="20.0" />
                </VBox.margin>
